Solved 2021 day 10 part 2

+++ src/main/kotlin/nl/mvdr/adventofcode/adventofcode2021/day10/ChunkType.kt
@@ -1,6 +1,6 @@
 package nl.mvdr.adventofcode.adventofcode2021.day10
 
-enum class ChunkType(val openBracket: Char, val closeBracket: Char, val syntaxErrorScore: Int, val completionScore: Int) {
+enum class ChunkType(val openBracket: Char, val closeBracket: Char, val syntaxErrorScore: Long, val completionScore: Long) {
     ROUND('(', ')', 3, 1),
     SQUARE('[', ']', 57, 2),
     CURLY('{', '}', 1197, 3),
